Add optional flag to clear output directory before package creation
- Added a new optional flag to clear the output directory before creating the package in the main function. This allows for purging the directory if specified by the user.

+++ create_package.py
@@ -30,6 +30,8 @@ import logging
 import collections
 import zipfile
 
+from typing import Optional
+
 # Name of addon
 #   - e.g. 'maya'
 ADDON_NAME = "shotgrid"
@@ -252,7 +254,12 @@ def create_server_package(output_dir, addon_output_dir, addon_version, log):
     log.info(f"Output package can be found: {output_path}")
 
 
-def main(output_dir=None, skip_zip=False, keep_sources=False):
+def main(
+    output_dir: Optional[str] = None,
+    skip_zip: bool = False,
+    keep_sources: bool = False,
+    clear_output_dir: bool = False
+):
     log = logging.getLogger("create_package")
     log.info("Start creating package")
 
@@ -268,9 +275,11 @@ def main(output_dir=None, skip_zip=False, keep_sources=False):
 
     addon_output_root = os.path.join(output_dir, ADDON_NAME)
     addon_output_dir = os.path.join(addon_output_root, addon_version)
-    if os.path.isdir(addon_output_root):
+
+    if os.path.isdir(addon_output_root) and clear_output_dir:
         log.info(f"Purging {addon_output_root}")
         shutil.rmtree(addon_output_root)
+
     os.makedirs(addon_output_dir)
 
     log.info(f"Preparing package for {ADDON_NAME}-{addon_version}")
@@ -311,6 +320,14 @@ if __name__ == "__main__":
         )
     )
     parser.add_argument(
+        "-c", "--clear-output-dir",
+        dest="clear_output_dir",
+        action="store_true",
+        help=(
+            "Clear output directory before package creation."
+        )
+    )
+    parser.add_argument(
         "-o", "--output",
         dest="output_dir",
         default=None,
@@ -321,4 +338,9 @@ if __name__ == "__main__":
     )
 
     args = parser.parse_args(sys.argv[1:])
-    main(args.output_dir, args.skip_zip, args.keep_sources)
+    main(
+        args.output_dir,
+        args.skip_zip,
+        args.keep_sources,
+        args.clear_output_dir
+    )
